Added DDSBASIC. Fixed an issue with gets which showed up while testing DDSBASIC. DDS reads in numbers over the top of its current command line, causing variable names to be scrubbed. it is a minimal BASIC after all.

+++ src/lib/conio/cgets.c
@@ -14,7 +14,7 @@ read start at str[2] and end with a null terminator. Thus, str must be at least 
 char* cgets(char* str)
 {
     str += 2;
-    eddie_line(str, str[-2], "\r", true);
+    eddie_line(str, (byte)str[-2], "\r", true);
     str[-1] = (char)strlen(str);
     return str;
 }
